Update ref-list.json to pass tests again
This might not be the cleanest way to fix this (we could manually add aliases instead) but I think most of these have never been used anyway. CURIE, e.g., was an alias for RDFA-CORE and html for HTML5 (see 2f5c235076246635cc7388e8364a13c5a9c181c7 for more about that).

This completes the work on tobie/specref#22