Give priority to href over edDraft in reverse lookup
Fix for #450.

Code that populated the reverse lookup table seemed to be willing to give
priority to `href` over `edDraft`, but it had 2 bugs:
1. it checked for `edDraft` in the properties of the output object not to
override a previous entry, but the entry that is inserted is
`normalizeUrl(edDraft)`, not `edDraft`
2. it did not add all `href` entries before looking at `edDraft`.

New test that makes sure that the returned ref is always the one that matches
`href` if there is one added.