Fix and update source URL in csswg.json
Last update to the script moved the `biblio.ref` URL to HTTPS but did not
update the source URL in `csswg.json`, which stuck to the HTTP version. This
meant that the script could no longer integrate updates from `biblio.ref`
(there haven't been any since then in practice).

This update switches the source URL to HTTPS, and updates it since it now is a
redirect to https://drafts.csswg.org/biblio.ref