Adjust URL and repo of Trusted Types (#720)
The Github repository was renamed last week. This updates the URL of the spec
and of the repository to reflect that change.

The spec should soon get published as a W3C First Public Working Draft so the
entry will likely soon need to be deleted from the `biblio.json` file.