Make file importable
This is coded assuming it is only ever imported by the `index.js` that runs it in the server with a process that cwd there, but it can be useful as a library. However, importing it as a library will fail to resolve `REFS_DIR`. This fix makes the code importable