Add browser-specs as new source of specs (#750)
The [w3c/browser-specs](https://github.com/w3c/browser-specs/) project contains
a curated list of specifications that compose the web platform. Most of the info
it contains comes from the W3C API (that info is already in Specref), and from
Specref itself.

That said, the list also contains additional specs that are not in Specref.
These are mostly "early" drafts developed in W3C community groups but not yet
published as Working Drafts, and WebGL extension specs.

With this update, Specref now retrieves specs from browser-specs that it does
not yet know anything about. This is useful for spec editors, because specs in
browser-specs get crawled and ingested in the cross-reference database of terms
used by Bikeshed and Respec, meaning that a spec editor can currently end up in
a situation where they can reference a term defined in a spec and yet are unable
to reference the spec itself.

The script runs first in `run-all` to drop specs from browser-specs as soon as
possible (removal from the browser-specs source should typically mean that the
spec has started to appear in another source, and it's going to be better to
rely on that other source directly).

The logic is based on `fetch-refs.js` with custom code to filter out specs that
are already in Specref, and preserve former entries that no longer exist in
browser-specs by creating aliases.

Via https://github.com/w3c/browser-specs/issues/975