Add URLPATTERN to whatwg.json and drop it from browser-specs.json
There seems to remain a transition hiccup when a WHATWG spec that was in
`browser-specs.json` starts appearing in `whatwg.json`. Last update to
`browser-specs.json` updated the `"URLPATTERN"` entry to flag it as an alias
of `"WHATWG-URLPATTERN"`, which seems correct as there was no `"URLPATTERN"`
entry in `whatwg.json`.

However, that update means that the `fetch-refs.js` script, applied to WHATWG
specs, now wants to create a `"URLPATTERN"` entry in `whatwg.json`. That
creates a duplicate, which prevents further scripts from running.

This update drops the alias entry from `browser-specs.json` and adds the right
entry to `whatwg.json` instead. This should allow builds to resume.