Refresh the list of RFCs published under httpwg.org (#766)
The list was no longer complete as the IETF HTTP WG published a few other RFCs.

Also, the https://httpwg.org/specs page is not such a good source for the list
because it fails to list a few specs, and also contains one RFC that is
actually not published under https://httpwg.org/specs. The comment explains how
to extract the list from the underlying GitHub repository instead.

In terms of updates, this will swap the URL of a few additional RFC entries
from `www.rfc-editor.org` to `httpwg.org/specs`, notably RFC9110, RFC9111,
RFC9112 and RFC9113. That's intended, the `httpwg.org/specs` versions are more
readable, as discussed in: https://github.com/w3c/browser-specs/issues/933