Fix browser-specs script for IETF drafts
Browser-specs contributes IETF drafts that have not yet been published as RFC
to Specref. A recent update made to browser-specs flag such entries as coming
from an "ietf" source (because the code gets the info from datatracker.ietf.org.

The browser-specs script needed to be updated in Specref to account for this
new source.