Update MSE info to handle levels 1, 2 and latest version (#769)
When MSE was published as a REC, it did not have levels. That changed. The
unversioned URL now targets Level 2, which is not the Rec.

This update renames the shortname of the Rec from `media-source` to
`media-source-1`, updates the URL for that level to
`https://www.w3.org/TR/media-source-1/`, and creates a `media-source` alias
to turn the `media-source` entry into a "current version" URL.

Specs that currently use `[[MEDIA-SOURCE]]` to reference MSE will now target
the level 2 WD and not the level 1 Rec, but then they were targeting the Rec
with the URL of WD, so the reference was clumsy in any case. All of these
specs are in the Media WG in any case (except HTML, but HTML references the
Editor's Draft, so nothing changes there...).

Note that there is no generic mechanism in Specref to handle such series. Some
discussion in #463.