Make the MSE alias target media-source-1 (#770)
Previous update missed the "MSE" alias that was hiding in `biblio.json`. The
alias targets `media-source`, which means it now targets `media-source-2`.
That's good but means that Specref can no longer resolve `MSE-YYYYMMDD` refs
to the dated level 1 versions.

In turn, that makes tests fail. +1 for the consistency tests. -1 for the
ability to alias names in different places!

This update creates the missing references as aliases explicitly, fixing tests.

I guess this shows that a more generic mechanism might be needed... but is not
trivial ;)

The new `w3c.json` file contains the new media-source entries (and the recently
published WoT specs)