Fix browser-specs entries for css-viewport and WASM GC proposal
Automatic updates currently crash because of browser-specs and for 2 reasons:

1. css-viewport was published to /TR under a different shortname. There remains
a chicken and egg issue there between the browser-specs list and the W3C list.
No easy way to handle that case, this manual intervention is needed for now to
create the alias ourselves.
2. The WASM GC proposal shortname changed in browser-specs from
`wasm-core-1-fork-gc` to `wasm-core-2-fork-gc`, and the entry in browser-specs
does not have the expected `formerNames`. That transition should have gone
smoothly if the update in browser-specs had been done correctly, meaning if a
`formerNames` property had been added. Will do that in browser-specs. Creating
the alias manually here in the meantime.