Force update of IETF references (#784)
Automatic updates are currently broken because of a chicken-and-egg problem
between browser-specs and ietf lists: browser-specs has a `DIGEST-HEADERS`
entry that used to be for an IETF draft, which now is `rfc9530`. The script
that updates the browser-specs wants to preserve the `DIGEST-HEADERS` entry,
turning it into an alias to `rfc9530`, but the script cannot find the RFC in
Specref because the IETF list does not yet have an entry for it.

The browser-specs script should probably be split into two parts: a part that
runs *before* the other scripts, and a part that runs *after* the other
scripts. In the meantime, refreshing the IETF list as done here should unblock
automatic updates.