Fix resource-timing entry in W3C list
The resource-timing spec has a clunky history because:
- the spec started without levels, so `resource-timing`;
- then moved to levels, with `resource-timing-1` redirecting to
`resource-timing-1` and then to `resource-timing-2`
- then dropped levels, so both `resource-timing-1` and `resource-timing-2` now
redirect to `resource-timing`

With the new redirects and related information in `tr.rdf`, the W3C script goes
berserk, destroys part of the history and creates an infinite loop, which
crashes tests and thus prevents further automatic updates.

To avoid this, this update makes `resource-timing` the main entry point with
the whole history, with levels as aliases of `resource-timing`.

This should unblock the generation and make the W3C script happy again.