Make encrypted-media an alias of encrypted-media-2 (#810)
The `encrypted-media` entry was an alias of `encrypted-media-1`, but the series
shortname should now redirect to `encrypted-media-2`.

Unfortunately, this means adding a bunch of aliases to preserve previous dated
entries. That's slightly ugly (although easy to prepare with a good text editor
that supports multi-line edits) but there's no real better way until Specref
understands series shortnames (tracked in #811).