Make auto updates run as a GitHub action
The script that updates the contents of Specref currently runs as a cron job on
a server. On top of costing a bit of money, this means that there is no easy
way to be notified about failures to update, and no easy way to access the logs
when an update fails. This update makes the `update-all` script as part of a
GitHub action instead.

Preliminaries:
- The following *variables* must be defined at the repo level:
  * `BOT_NAME`, set to the GitHub account of the bot (typically: specworker)
  * `BOT_EMAIL`, set to the email of the bot
- The following *secret* must be defined at the repo level:
  * `BOT_TOKEN`, set to a "Personal access token (classic)" with `public_repo`
  scope in the GitHub account of the bot

(The bot account must of course have write permission on the repository. That
is already the case for specworker).

The script runs every hour. It may also be triggered manually from the user
interface on GitHub.