Fix SMIL20 aliases and overwrite rules for media-source (#825)
* Fix SMIL20 aliases and overwrite rules for media-source

I'm working with @deniak on improving the data in the W3C API so that Specref
may more easily switch to the W3C API. One rippling effect is that this means
that some of the hacks done to get the right entries are no longer needed (and
now create duplicates). Adjusting the rules accordingly.

* Add updated w3c.json

(That's the result of running the W3C script. Needed for tests to pass)