Create real entries for 1999-05-06 versions of xmlschema
The `xmlschema-1-1` and `xmlschema-2-2` entries in Specref are shortcuts for
`xmlschema-1-19990506` and `xmlschema-2-19990506`, which do not exist in
`tr.rdf` for some reason. This update creates the missing versions and makes
the shortcuts aliases of these entries.

(This update is needed to switch to the W3C API as source for the data, as
the W3C API does have the 1999-05-06 versions)