Drop broken links to a few W3C Editor's Drafts
The previously known Editor's Drafts of a few legacy and now retired W3C
specifications now return a 404 or redirect to a page that is not a spec. These
specs have a newer version somewhere, but it does not make sense to target the
new ED from the retired entry.