Drop overwrites / Force refresh of user-timing entries.
The overwrites will only become needed once the related entries get updated
and that will only happen once we run a full refresh. In the meantime, they
create missing entries.

The User Timing specs have been going back and forth between shortnames,
which led to some entries incorrectly associated with `user-timing-1`, while
last publications did not appear under `user-timing-3`. This update performs
a full refresh of the entries and adds a few aliases for continuity (these
entries should not have existed at all but since they do, we need to keep
them around). Main entries are now `user-timing-1`, `user-timing-2` and
`user-timing`, with `user-timing-3` defined as an alias of `user-timing`,
consistent with the info in the W3C API and with how the Web Performance WG wanted to proceed with the third level.

The entries for `vocab-dcat`, `xmlenc-core`, `wot-architecture` and
`wot-thing-description` will need to be manually fixed as well, but that can
wait until we run a full sync.