Bump Node version in package.json, re-genereate package-lock.json (#820)
The `package-lock.json` file was generated with a now obsolete version of npm
and any attempt to run `npm install` with a more recent version of npm starts
with converting the file to the new JSON format. This update does this so
that installing the code (or in a job) does not require creating changes to a
repo file.

The `package.json` file also referenced a now obsolete and no longer supported
version of Node.js (v14). Using the opportunity to bump the versions to
something more current (>v20.14 for Node, >v10.7 for npm). Any sufficiently
recent install of Node.js will match these constraints, and this will allow
using more advanced Node.js features (including `fetch` to replace the
dependency on `request` down the road).