Support spec series and retrieve them for W3C specs
This adds support for specification series, as discussed in #811. Actual
changes to the core logic are minimal and should be fully transparent for
consumers: a new `isSeriesAlias` boolean flag can now be added to an alias to
signal that it is a "series alias". This alias gets processed by bibref exactly
as any other alias, except that bibref does not create version entries when the
flag is set.

In other words, given:

```
{
  "css-fonts": {
    "aliasOf": "css-fonts-4",
    "isSeriesAlias": true
  }
}
```

`css-fonts` can then be used to reference `css-fonts-4` but, as opposed to
a regular alias, `css-fonts-20240201` cannot be used to reference
`css-fonts-4-20240201`. This is on purpose as it allows to update `css-fonts`
later on to target `css-fonts-5` when it becomes the "current" specification.

The W3C script now leverages the W3C API to retrieve the list of specification
series and create series aliases when needed and possible.

Consumers won't even know that they are dealing with a series, but it's not
clear that they need to know in any case, and we can make that more explicit
later on if needed.

Fixes #811.