Make `controller-document` an alias of new `cid` entries
The shortname of the Controller Document specification was changed from
`controller-document` to `cid`. The W3C script attempts to create `cid` entries
but that creates URL duplicates, making tests fail.

This update adds the `cid` entries and makes `controller-document` an alias of
them.

Note: such renames are reported by the script but not handled automatically
for now because there are a few other legacy cases where automatic handling
would actually break things.