Make CSS21 an alias of CSS2
CSS2 used to be an alias of CSS21 in both Specref and the W3C API, but the
relationship got swapped in the W3C API a couple of days ago as the final
intent is that revision 1 and revision 2 of CSS Level 2 end up being known as
simply CSS2.

This update swaps the alias direction in Specref as well to align with the W3C
API. That update is actually needed to fix the build because the W3C API now
reports the `obsoletes` property for CSS2, which created an incorrect alias
entry in the JSON file.