Force deletion of generic "draft" and "standard" WG21 entries
Two new entries got added to the WG21 JSON source recently: "draft" and
"standard". Both entries have the same link, which (rightfully) makes Specref
choke. On top of that, these entries have very generic names and don't make
sense in Specref.

This completes the list of overwrites for WG21 to delete the "draft" and
"standard" entries (and their prefixed versions that the code automatically
creates).